New method Functor::fmap_same
A new method `Functor::fmap_same` has been added to simplify handling
cases where the closure passed to `fmap` returns the same type as its
argument (e.g. when mapping a `Vec<T>` to a `Vec<T>`). This is an
alternative to the helper methods `FunctorSelf::from_mapped` and
`FunctorSelf::into_mapped`.