Added applicative functors
Trait `Applicative` has been added to describe applicative functors.
Functions `applicative_fmap` and `monad_apply` serve as possible default
implementations for `Functor::fmap` and `Applicative::apply`, based on
the `Applicative` and `Monad` implementation, respectively.

Some of the documentation for `Monad` was moved from the top-level
module to the `Monad` trait.